backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic>=2.1.0,<3.0.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx<=0.24.1
stripe==5.4.0
supabase==1.0.3
openai>=1.0.0
anthropic>=0.8.0
pytest==7.3.1
pytest-asyncio==0.21.0
black==23.3.0
isort==5.12.0
gotrue>=1.0.3
prometheus-client==0.17.1
pydantic-settings==2.0.3
email-validator==2.1.0.post1
python-dotenv==1.0.0
redis>=5.0.0
